src/models/stan_model.py
# Import necessary libraries for different Stan interfaces
import pystan
import cmdstanpy
import httpstan
import bridgestan


class StanModel:
    def __init__(self, model_code, interface='pystan'):
        self.model_code = model_code
        self.interface = interface

        if interface == 'pystan':
            self.model = pystan.StanModel(model_code=self.model_code)
        elif interface == 'cmdstanpy':
            self.model = cmdstanpy.CmdStanModel(stan_file=self.model_code)
        elif interface == 'bridgestan':
            # Initialize BridgeStan model, assuming a similar API to other Stan interfaces
            self.model = bridgestan.StanModel(model_code=self.model_code)
        elif interface == 'httpstan':
            # Initialize httpstan model, this might require an async call or a server setup
            self.model = httpstan.build_model(self.model_code)
        else:
            raise ValueError(f"Unsupported interface: {interface}")

    def fit(self, data, **kwargs):
        if self.interface == 'pystan':
            fit = self.model.sampling(data=data, **kwargs)
        elif self.interface == 'cmdstanpy':
            fit = self.model.sample(data=data, **kwargs)
        elif self.interface == 'bridgestan':
            # Fit the model using BridgeStan's fitting method
            fit = self.model.fit(data=data, **kwargs)
        elif self.interface == 'httpstan':
            # Fit the model using httpstan's fitting method, may involve an HTTP request
            fit = self.model.fit(data=data, **kwargs)
        else:
            raise ValueError(f"Unsupported interface: {self.interface}")

        return fit

    # ...additional methods for diagnostics, predictions, etc...
